+++ VideoCore/src/main/cpp/Parser/ParseRTP.h
@@ -16,11 +16,11 @@ class ParseRTP{
 public:
     ParseRTP(NALU_DATA_CALLBACK cb);
 public:
-    void parseData(const uint8_t* data,const int data_length);
+    void parseData(const uint8_t* data,const size_t data_length);
     void reset();
 private:
     const NALU_DATA_CALLBACK cb;
     uint8_t nalu_data[NALU::NALU_MAXLEN];
-    int nalu_data_length=0;
+    unsigned int nalu_data_length=0;
 };
 #endif //LIVE_VIDEO_10MS_ANDROID_PARSERTP_H
